requirements.txt
llama-cpp-python
requests==2.26.0